requirements.txt
streamlit
pandas
xlsxwriter
google-generativeai
Pillow
gspread
google-auth
requests
toml